Add transaction to wallet for compatibility RPCs.
The Xaya compatibility RPC interface provides name_update and name_register
methods that not only construct the transaction but also broadcast them.
This change also explicitly calls "addtransaction" for them, which ensures
that they can be immediately seen by the wallet (not only after a roundtrip
from the server).  With this, we can issue multiple updates to the same
name quickly after each other.